Fetch for the robots users
