--- src/App.js
@@ -1,40 +1,0 @@
-import React from 'react';
-import CardList from './CardList';
-import SearchBox from './SearchBox';
-import { robots } from './robots';
-import './App.css'
-
-class App extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            robots,
-            searchfield: ''
-        }
-    }
-
-    onSearchChange = (event) => {
-        this.setState({
-            searchfield: event.target.value
-        })
-
-    }
-
-    render() {
-        const filteredRobots = this.state.robots.filter((robots) => {
-            return robots.name
-                .toLowerCase()
-                .includes(this.state.searchfield.toLowerCase());
-        })
-
-        return (
-            <div className='tc'>
-                <h1 className="f2">RoboFriends</h1>
-                <SearchBox searchChange={this.onSearchChange} />
-                <CardList robots={filteredRobots} />
-            </div>
-        );
-    }
-}
-
-export default App;
